// server/src/utils/enum.ts
export enum statusMessage {
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
  USER_ALREADY_REGISTERED = 'User is already registered',
  USER_NOT_REGISTERED = 'User is not registered',
  INCORRECT_PASSWORD = 'Incorrect Password',
  TOKEN_NOT_FOUND = 'Token was not found',
  PERMISSIONS_DID_NOT_MATCH = 'Permissions did not match',
  CHAT_COMPLETION_REQUEST_MESSAGE = 'Chat complete request message error',
  INTERNAL_SERVER_ERROR = 'Internal Server Error',
  TOKEN_EXPIRED_INVALID = 'Token is expired or invalid'
}